requirements.txt
opencv-python>=4.5.0
pytesseract>=0.3.8
pyautogui>=0.9.53
numpy>=1.19.5
plyer>=2.0.0
Pillow>=8.4.0 